soc/intel/common: move common memmap functionality from skl,icl,cnl,apl
This moves common memmap functionality from skl,icl,cnl,apl to the common tree.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/36410
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
