+++ src/soc/intel/common/block/systemagent/Makefile.inc
@@ -3,3 +3,6 @@ romstage-$(CONFIG_SOC_INTEL_COMMON_BLOCK_SA) += systemagent_early.c
 ramstage-$(CONFIG_SOC_INTEL_COMMON_BLOCK_SA) += systemagent_early.c
 postcar-$(CONFIG_SOC_INTEL_COMMON_BLOCK_SA) += systemagent_early.c
 ramstage-$(CONFIG_SOC_INTEL_COMMON_BLOCK_SA) += systemagent.c
+romstage-$(CONFIG_SOC_INTEL_COMMON_BLOCK_SA) += memmap.c
+ramstage-$(CONFIG_SOC_INTEL_COMMON_BLOCK_SA) += memmap.c
+postcar-$(CONFIG_SOC_INTEL_COMMON_BLOCK_SA) += memmap.c
